+++ mne/tests/test_evoked.py
@@ -14,9 +14,9 @@ from numpy.testing import (assert_array_almost_equal, assert_equal,
                            assert_array_equal, assert_allclose)
 from nose.tools import assert_true, assert_raises, assert_not_equal
 
-from mne import equalize_channels, pick_types, read_evokeds, write_evokeds
-from mne import grand_average
-from mne.evoked import _get_peak, EvokedArray, grand_average
+from mne import (equalize_channels, pick_types, read_evokeds, write_evokeds,
+                 grand_average, combine_evoked)
+from mne.evoked import _get_peak, EvokedArray, merge_evoked
 from mne.epochs import EpochsArray
 
 from mne.utils import _TempDir, requires_pandas, requires_nitime
@@ -333,9 +333,45 @@ def test_equalize_channels():
         assert_equal(ch_names, e.ch_names)
 
 
-def test_grand_average():
-    """Test grand average
+def test_evoked_arithmetic():
+    """Test evoked arithmetic
     """
+    ev = read_evokeds(fname, condition=0)
+    ev1 = EvokedArray(np.ones_like(ev.data), ev.info, ev.times[0], nave=20)
+    ev2 = EvokedArray(-np.ones_like(ev.data), ev.info, ev.times[0], nave=10)
+    
+    # combine_evoked([ev1, ev2]) should be the same as ev1 + ev2:
+    # data should be added according to their `nave` weights
+    # nave = ev1.nave + ev2.nave
+    ev = ev1 + ev2
+    assert_equal(ev.nave, ev1.nave + ev2.nave)
+    assert_allclose(ev.data, 1. / 3. * np.ones_like(ev.data))
+    ev = ev1 - ev2
+    assert_equal(ev.nave, ev1.nave + ev2.nave)
+    assert_allclose(ev.data, np.ones_like(ev1.data))
+    with warnings.catch_warnings(record=True) as w:
+        warnings.simplefilter('always')
+        ev = merge_evoked([ev1, ev2])
+    assert_true(len(w) >= 1)
+    assert_allclose(ev.data, 1. / 3. * np.ones_like(ev.data))
+
+    # equal weighting
+    ev = combine_evoked([ev1, ev2], weights='equal')
+    assert_allclose(ev.data, np.zeros_like(ev1.data))
+
+    # combine_evoked([ev1, ev2], weights=[1, 0]) should yield the same as ev1
+    ev = combine_evoked([ev1, ev2], weights=[1, 0])
+    assert_equal(ev.nave, ev1.nave)
+    assert_allclose(ev.data, ev1.data)
+
+    # simple subtraction (like in oddball)
+    ev = combine_evoked([ev1, ev2], weights=[1, -1])
+    assert_allclose(ev.data, 2 * np.ones_like(ev1.data))
+
+    assert_raises(ValueError, combine_evoked, [ev1, ev2], weights='foo')
+    assert_raises(ValueError, combine_evoked, [ev1, ev2], weights=[1])
+
+    # grand average
     evoked1, evoked2 = read_evokeds(fname, condition=[0, 1], proj=True)
     ch_names = evoked1.ch_names[2:]
     evoked1.info['bads'] = ['EEG 008']  # test interpolation